demo/requirements.txt
Click==7.0
deepmerge==0.0.5
Faker==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-dateutil==2.8.0
six==1.12.0
text-unidecode==1.2
Werkzeug==0.14.1
